+++ starflut/CHANGELOG.md
@@ -1,3 +1,22 @@
+## 0.6.0
+
+* release at 2019/08/01
+* starcore upgrade to 3.5.0
+
+* Newly added API Interface:
+> StarServiceClass:
+  >> [allObject](#)            : Future<List> allObject() async 
+  >> [restfulCall](#)          : Future<List> restfulCall (String url, String opCode, String jsonString) async
+
+> StarParaPkg:
+  >> [equals](#)               : Future<bool> equals (StarParaPkgClass srcParaPkg) async
+  >> [v](#)                    : Future<String> get ValueStr async
+
+> StarObject :
+  >> [instNumber](#)           : Future<int> instNumber () async
+  >> [jsonCall](#)             : Future<String> jsonCall (String jsonString) async
+
+ 
 ## 0.5.2
 
 * release at 2018/08/03